+++ Veg Out Taiwan/Controller/Restaurant/UserCommentWallViewController.swift
@@ -77,7 +77,7 @@ extension UserCommentWallViewController: UICollectionViewDelegateFlowLayout {
     
     func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
           
-          return CGSize(width: (collectionView.frame.width - 40) / 3, height: collectionView.frame.width / 3)
+          return CGSize(width: Int((collectionView.frame.width - 40) / 3), height: Int(collectionView.frame.width / 3))
       }
       
       func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, insetForSectionAt section: Int) -> UIEdgeInsets {
